// src/api/dezswap/models/getDashboardChartTokensAddressTypeDuration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */

export type GetDashboardChartTokensAddressTypeDuration =
  (typeof GetDashboardChartTokensAddressTypeDuration)[keyof typeof GetDashboardChartTokensAddressTypeDuration];

export const GetDashboardChartTokensAddressTypeDuration = {
  year: "year",
  quarter: "quarter",
  month: "month",
} as const;